docker/07-docker-compose-avanzado/app/requirements.txt
Flask==2.3.3
redis==5.0.1
psycopg2-binary==2.9.9

